backend/requirements.txt
openai>=1.40.0
pydantic>=2.8.0
pypandoc-binary>=1.12
google-genai>=0.3.0
numpy>=1.26
scipy>=1.12
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
